// frontend/src/lib/proto/beautiulmind.ts
export interface Card {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * @generated from protobuf field: string deck_id = 2
     */
    deckId: string;
    /**
     * @generated from protobuf field: string concept_id = 3
     */
    conceptId: string;
    /**
     * @generated from protobuf field: string title = 4
     */
    title: string;
    /**
     * @generated from protobuf field: string content = 5
     */
    content: string;
}
/**
 * @generated from protobuf message beautifulmind.Concept
 */
export interface Concept {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
}
/**
 * @generated from protobuf message beautifulmind.JourneyStep
 */
export interface JourneyStep {
    /**
     * @generated from protobuf field: beautifulmind.Card card = 1
     */
    card?: Card;
    /**
     * @generated from protobuf field: bool answered_correctly = 2
     */
    answeredCorrectly: boolean;
}
/**
 * represents the entire state of a game session
 *
 * @generated from protobuf message beautifulmind.SessionState
 */
export interface SessionState {
    /**
     * @generated from protobuf field: beautifulmind.Card thesis = 1
     */
    thesis?: Card;
    /**
     * @generated from protobuf field: repeated beautifulmind.Card evidence_deck = 2
     */
    evidenceDeck: Card[];
    /**
     * @generated from protobuf field: int32 current_step = 3
     */
    currentStep: number;
    /**
     * @generated from protobuf field: repeated beautifulmind.JourneyStep journey_log = 4
     */
    journeyLog: JourneyStep[];
    /**
     * @generated from protobuf field: bool is_evidence_revealed = 5
     */
    isEvidenceRevealed: boolean;
    /**
     * @generated from protobuf field: beautifulmind.Card current_card = 6
     */
    currentCard?: Card;
    /**
     * @generated from protobuf field: map<string, beautifulmind.Concept> concepts = 7
     */
    concepts: {
        [key: string]: Concept;
    };
}
// ...
class Card$Type extends MessageType<Card> {
    constructor() {
        super("beautifulmind.Card", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "deck_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "concept_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "title", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "content", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<Card>): Card {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.deckId = "";
        message.conceptId = "";
        message.title = "";
        message.content = "";
        if (value !== undefined)
            reflectionMergePartial<Card>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Card): Card {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string deck_id */ 2:
                    message.deckId = reader.string();
                    break;
                case /* string concept_id */ 3:
                    message.conceptId = reader.string();
                    break;
                case /* string title */ 4:
                    message.title = reader.string();
                    break;
                case /* string content */ 5:
                    message.content = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Card, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string deck_id = 2; */
        if (message.deckId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.deckId);
        /* string concept_id = 3; */
        if (message.conceptId !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.conceptId);
        /* string title = 4; */
        if (message.title !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.title);
        /* string content = 5; */
        if (message.content !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.content);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message beautifulmind.Card
 */
export const Card = new Card$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Concept$Type extends MessageType<Concept> {
    constructor() {
        super("beautifulmind.Concept", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<Concept>): Concept {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.name = "";
        if (value !== undefined)
            reflectionMergePartial<Concept>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Concept): Concept {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string name */ 2:
                    message.name = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Concept, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string name = 2; */
        if (message.name !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.name);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message beautifulmind.Concept
 */
export const Concept = new Concept$Type();
// @generated message type with reflection information, may provide speed optimized methods
class JourneyStep$Type extends MessageType<JourneyStep> {
    constructor() {
        super("beautifulmind.JourneyStep", [
            { no: 1, name: "card", kind: "message", T: () => Card },
            { no: 2, name: "answered_correctly", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<JourneyStep>): JourneyStep {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.answeredCorrectly = false;
        if (value !== undefined)
            reflectionMergePartial<JourneyStep>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: JourneyStep): JourneyStep {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* beautifulmind.Card card */ 1:
                    message.card = Card.internalBinaryRead(reader, reader.uint32(), options, message.card);
                    break;
                case /* bool answered_correctly */ 2:
                    message.answeredCorrectly = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: JourneyStep, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* beautifulmind.Card card = 1; */
        if (message.card)
            Card.internalBinaryWrite(message.card, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* bool answered_correctly = 2; */
        if (message.answeredCorrectly !== false)
            writer.tag(2, WireType.Varint).bool(message.answeredCorrectly);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message beautifulmind.JourneyStep
 */
export const JourneyStep = new JourneyStep$Type();
// ...
export const SessionState = new SessionState$Type();